requirements.txt
keras==2.12.0
keras_facenet==0.3.2
matplotlib==3.5.1
nltk==3.8.1
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
pandas==1.4.2
Pillow==9.5.0
Pillow==9.5.0
Sastrawi==1.0.1
Sastrawi==1.0.1
scikit_learn==1.1.1
scipy==1.8.1
tensorflow==2.12.0
tensorflow_intel==2.12.0
tkcalendar==1.6.1
